# Improve performance of the admin page