--- a/rest_framework_tracking/admin.py
+++ b/rest_framework_tracking/admin.py
@@ -6,10 +6,10 @@ from .models import APIRequestLog
 class APIRequestLogAdmin(admin.ModelAdmin):
     date_hierarchy = 'requested_at'
     list_display = ('id', 'requested_at', 'response_ms', 'status_code',
-                    'user', 'method',
+                    'user', 'view_method',
                     'path', 'remote_addr', 'host',
                     'query_params')
-    list_filter = ('method', 'status_code')
+    list_filter = ('view_method', 'status_code')
     search_fields = ('path', 'user__email',)
     raw_id_fields = ('user', )