--- a/rest_framework_tracking/base_models.py
+++ b/rest_framework_tracking/base_models.py
@@ -43,7 +43,7 @@ class BaseAPIRequestLog(models.Model):
     data = models.TextField(null=True, blank=True)
     response = models.TextField(null=True, blank=True)
     errors = models.TextField(null=True, blank=True)
-    status_code = models.PositiveIntegerField(null=True, blank=True)
+    status_code = models.PositiveIntegerField(null=True, blank=True, db_index=True)
     objects = PrefetchUserManager()
 
     class Meta: