The admin page runs into performance issues when the `rest_framework_tracking_apirequestlog` table gets a decent number of records. The list filters being applied to the `method` and `status_code` fields are slow because they run a `distinct` query on those fields, which results in a slow sequential scan because those fields are not indexed.

This pull request:
1) switches to the `view_method` field instead of `method` because `view_method` already has an index, and
2) adds an index to the `status_code` field

For example:

BEFORE
Query:
```
EXPLAIN ANALYZE SELECT DISTINCT "rest_framework_tracking_apirequestlog"."method" FROM "rest_framework_tracking_apirequestlog" ORDER BY "rest_framework_tracking_apirequestlog"."method" ASC;
```
Result:
```
 Sort  (cost=1260443.38..1260443.39 rows=4 width=4) (actual time=43682.517..43682.521 rows=6 loops=1)
   Sort Key: method
   Sort Method: quicksort  Memory: 25kB
   ->  HashAggregate  (cost=1260443.30..1260443.34 rows=4 width=4) (actual time=43682.467..43682.472 rows=6 loops=1)
         Group Key: method
         ->  Seq Scan on rest_framework_tracking_apirequestlog  (cost=0.00..1259094.84 rows=539384 width=4) (actual time=16.158..43510.179 rows=529529 loops=1)
 Planning Time: 0.150 ms
 Execution Time: 43682.571 ms
```

AFTER
Query:
```
EXPLAIN ANALYZE SELECT DISTINCT "rest_framework_tracking_apirequestlog"."view_method" FROM "rest_framework_tracking_apirequestlog" ORDER BY "rest_framework_tracking_apirequestlog"."view_method" ASC;
```
Result:
```
 Unique  (cost=0.42..32249.64 rows=4 width=4) (actual time=0.583..1581.838 rows=6 loops=1)
   ->  Index Only Scan using rest_framework_tracking_apirequestlog_view_method_dd790881 on rest_framework_tracking_apirequestlog  (cost=0.42..30901.18 rows=539384 width=4) (actual time=0.581..1521.871 rows=529529 loops=1)
         Heap Fetches: 64
 Planning Time: 0.160 ms
 Execution Time: 1581.867 ms
``