@lingster 2020-11-24
@anatolzak : apologises for the delay, have been busy with other things, this PR is now merged. Thanks for your contribution!